bgpv2: make status update interval configurable
Introducing bgp-status-update-interval config map option to set how
frequently agent will update status field of CiliumBGPNodeConfig
resource. Depending on the expected updates to BGP and number of nodes
in the cluster, user may want to tune this option so as to not overload
API server. By default, it is 5s.

Signed-off-by: harsimran pabla <hpabla@isovalent.com>